+++ test/test322.js
@@ -13,7 +13,7 @@ describe('Test 322 UNION TEST', function() {
     done();
   });
 
-  it('2. CREATE GRAPH',function(done){
+  it('2. UNION ALL',function(done){
     var data = [{a:1},{a:2},{a:2}, {b:2}];
     var res = alasql('SELECT a FROM $0 WHERE NOT a IS NULL \
       UNION ALL CORRESPONDING SELECT b FROM $0 WHERE NOT b IS NULL',[data]);
@@ -29,8 +29,8 @@ describe('Test 322 UNION TEST', function() {
     assert.deepEqual(res,[ { a: 1 }, { a: 2 }, { a: 2 }, { a: 2 } ]);
 
     var res = alasql('SELECT a FROM $0 WHERE NOT a IS NULL \
-      UNION SELECT b FROM $0 WHERE NOT b IS NULL',[data]);
-    assert.deepEqual(res,[ { a: 1 }, { a: 2 }, { a: 2 }]); // To be checked
+      UNION SELECT b FROM $0 WHERE NOT b IS NULL ORDER BY a',[data]);
+    assert.deepEqual(res,[ { a: 1 }, { a: 2 }]); // To be checked
 // or 1,2,2
 
 //    console.log(res);
@@ -42,7 +42,7 @@ describe('Test 322 UNION TEST', function() {
     var data = [{a:1},{a:2},{a:2}, {b:2}];
 
     var res = alasql('SEARCH UNION(/a,/b) FROM ?',[data]);
-    assert.deepEqual(res,[1,2,2]);
+    assert.deepEqual(res,[1,2]);
     done();
   });